Add grammar tests for label, goto, and forever
I also removed the loop until command as loop was then a keyword which
could not be used as a parameter to any functions and in particular the
playSound function. Considering this I also removed the until command as
I think it is not any easier to learn this than a while loop. I did keep
forever though.